+++ changelog.md
@@ -1,6 +1,8 @@
 # 1.4.1
 - Added garage preview support
 - Added profile icon mod support and compatibility
+- Added rainbow effect to profile icons and fixed glow in garage
+
 
 # 1.4.0
 - Fixed the flashPlayer wich was making the icon white permanently